move 2 modules from appledict/ to appledict_bin/
