+++ pyglossary/plugins/appledict_bin/appledict_file_tools.py
